refactor: updated Icons, file references for consistency and improved structure.
