+++ .gitignore
@@ -45,3 +45,6 @@ pnpmfile.cson
 # generated css files
 **/scss/**/*.css
 .vercel
+
+**/*.css
+**/*.css.map
