+++ src/styles/components/common/Footer.module.scss
@@ -1,4 +1,4 @@
-@use "../../abstracts/typography.module.css" as *;
+@use "../../abstracts/typography.module" as *;
 @use "../../partials/variables" as *;
 
 .footer {
