+++ src/utils/Icons.tsx
@@ -18,9 +18,9 @@ const Icons = {
   ShoppingCart: createIcon("hugeicons:shopping-cart-01"),
   GridIcon: createIcon("ci:grid-big-round"),
   Warranty: createIcon("ph:seal-check-light"),
-  Trophy: createIcon("fluent:trophy-32-regular"),
-  Delivery: createIcon("icon-park-twotone:delivery"),
-  CustomerService: createIcon("bx:support"),
+  Trophy: createIcon("mdi-light:trophy"),
+  Delivery: createIcon("material-symbols-light:delivery-truck-speed-outline"),
+  CustomerService: createIcon("ph:headset-light"),
   Share: createIcon("material-symbols:share"),
   ViewList: createIcon("bi:view-list"),
   XFill: createIcon("flowbite:close-circle-solid"),
